fix(gguf): the qwen3 arm cannot read any published Qwen3 GGUF, and fix the MSVC narrowing that killed all three Windows lanes
Two independent defects found by running the merged tree rather than reading it.

1. The qwen3 GGUF arm reads a tensor name no real file contains.

Upstream 3c2fcca33 landed qwen3 GGUF support on 2026-09-23 and this tree merged
it the same day. On a Jetson Orin Nano, against a real Qwen3-0.6B-BF16.gguf
(311 tensors, general.architecture = qwen3):

  vllm_engine_load: gguf: no tensor named "blk.0.post_attention_norm.weight"

The file names it blk.0.ffn_norm.weight, which is llama.cpp's converter output
and therefore what every Qwen3 GGUF on the hub contains. post_attention_norm is
this tree's OWN renaming, applied by its own converters
(muse_glimmer_gguf_weights.cpp:568 maps post_attention_layernorm.weight to it).

The identification is structural, not a guess: blk.0 in that file carries
exactly eleven tensors, this loader asks for exactly eleven, the two sets differ
in this ONE element, and the three non-block names (token_embd, output_norm,
output) match exactly. A Qwen3 layer has exactly two layernorms and attn_norm is
already claimed as the input one, so ffn_norm is the only tensor
post_attention_layernorm can be. The loader now prefers ffn_norm and falls back
to post_attention_norm, so tree-converted files keep working.

WHY THE TEST WAS GREEN. test_qwen3_gguf_weights.cpp builds a SYNTHETIC GGUF and
wrote whichever name the loader read (its line 109), so it agreed with the
loader by construction and could not see that no published file loads. That is
this project's own rule arriving from the other side: a control that shares the
bug is not a control. BuildGguf now takes the spelling as a parameter, defaults
to the stock one, and a new case loads both.

CONSEQUENCE FOR THE RELEASE NOTES, and it corrects something already written:
docs/vllm-public-upstream-merge-2026-09-23.md said this merge removed the
container asymmetry so both engines could read one file. It does not, without
this fix -- the arm parses the config and then refuses the weights.

2. gemma3.cpp:579 narrows int64_t to int, which MSVC treats as an error.

  gemma3.cpp(579,34): error C2220: the following warning is treated as an error
  warning C4244: '=': conversion from 'const _Ty' to 'int', possible loss of data

s.columns is int (line 535) and feeds block_table_num_cols, which is int too, so
the narrowing belongs to the destination rather than to this line -- but it has
to be spelled, because GCC and Clang say nothing by default and MSVC raises
C4244 at /W3. That asymmetry is the whole story of run 35855557148: all three
Linux lanes green, all three Windows lanes dead on this one line. Verified by
building the vllm library under MSVC 14.44.

Not fixed here, and deliberately: test_qwen3_gguf_weights.cpp has its own
pre-existing C4244 sites and does not compile under MSVC. It is not in
build-windows-release.ps1's target list, so it does not gate the release, and
widening this commit into the test suite's MSVC port during a release freeze
would be scope this change has not earned.

+++ src/vllm/model_executor/models/gemma3.cpp
@@ -576,7 +576,15 @@ ForwardLogits Gemma3DecodeGraph::Step(const ModelForwardInput& in) {
   }
   if (new_inputs) {
     s.Reset();
-    s.columns = std::max<int64_t>(in.attn_meta.block_table_num_cols, in.attn_kv.front().num_blocks);
+    // `s.columns` is int (line 535) and feeds `block_table_num_cols`, which is
+    // int too, so the narrowing is inherent to the destination rather than a
+    // loss this line invents -- but it must be SPELLED, because MSVC raises
+    // C4244 at /W3 and this tree compiles warnings as errors. GCC and Clang
+    // say nothing by default, which is why all three Linux lanes were green
+    // while all three Windows lanes died here (run 35855557148). The cast
+    // matches what line 583 already does two lines below.
+    s.columns = static_cast<int>(std::max<int64_t>(
+        in.attn_meta.block_table_num_cols, in.attn_kv.front().num_blocks));
     s.cache = in.attn_kv;
     s.metadata = in.attn_meta;
     s.metadata.block_table_num_cols = s.columns;
